Overhaul handling of value normalization
The serialization is now more compact, with entries for lists and maps
not having to repeat the constraints and descriptions, which are now placed
in the root value.

Maps have key constraints in addition to value constraints. The data is
serialized as a list rather than a map in order to support complex keys
(function calls) in formats that do not support non-string hashtable keys.