Major refactor to documentation and profiles
The Kubernetes profile has been removed, as it is now managed within the
Turandot project. This means that all examples based on Kubernetes have
also been removed. Likewise, the OpenStack and BPMN profile are no longer
built-in.

The QUICKSTART guide is now much more substantial and focuses on using
all the tools.

Other changes:

* Major improvements to puccini-csar
* Generalize 'tosca.lib.traversal' JavaScript
* Begin work on "diff" tool for Clout
* Fix for Ruby wrapper build