Refactoring of value normalization and coercion
This is a major refactoring. The problem was that the previous
implementation was no recursive enough and could not handle constraints
and other metadata within schemas and other sub-types of maps. The
solution required a complete rewrite of Clout $meta to be truly
recursive. This affected parsing, normalization, and coercion.

Also included:

* Upgrade to Go 1.19
* Upgrade many dependencies
* Support for MessagePack (alternative to CBOR)
* A lot of code cleanup
* Prepare for supporting TOSCA 2.0 function-based validations (to
  replace constraints)